# public/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: 'Takebook API'
  contact:
    email: support@example.com
  version: 1.0.0
paths:
...
  /books:
    get:
      tags:
        - books
      summary: 'Lista todos os livros'
      operationId: GetBooks
      parameters:
        -
          name: title
          in: query
          description: 'Nome do livro'
          required: false
          schema:
            type: string
        -
          name: author
          in: query
          description: 'Autor do livro'
          required: false
          schema:
            type: string
        -
          name: user_id
          in: query
          description: 'ID do anunciante'
          required: false
          schema:
            type: string
        -
          name: includes
          in: query
          description: 'Faz o include das relações '
          required: false
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: ...
    post:
      tags:
        - books
      summary: 'Cria um novo livro'
      operationId: PostBooks
      parameters:
        -
          name: title
          in: query
          description: 'Nome do livro'
          required: true
          schema:
            type: string
        -
          name: author
          in: query
          description: 'Autor do livro'
          required: true
          schema:
            type: string
        -
          name: description
          in: query
          description: 'Descrição do livro'
          required: true
          schema:
            type: string
        -
          name: price
          in: query
          description: 'Preço do livro'
          required: true
          schema:
            type: string
        -
          name: categories
          in: query
          description: 'IDs das categorias'
          required: true
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: ...
      security:
        -
          apiToken: []
  '/books/{id}':
    get:
      tags:
        - books
      summary: 'Lista um livro'
      operationId: GetBook
      parameters:
        -
          name: id
          in: path
          description: 'ID do livro'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ...
      security:
        -
          apiToken: []
    put:
      tags:
        - books
      summary: 'Atualiza um livro'
      operationId: PutBooks
      parameters:
        -
          name: id
          in: path
          description: 'ID do livro'
          required: true
          schema:
            type: integer
        -
          name: title
          in: query
          description: 'Nome do livro'
          required: false
          schema:
            type: string
        -
          name: author
          in: query
          description: 'Autor do livro'
          required: false
          schema:
            type: string
        -
          name: description
          in: query
          description: 'Descrição do livro'
          required: false
          schema:
            type: string
        -
          name: price
          in: query
          description: 'Preço do livro'
          required: false
          schema:
            type: string
        -
          name: categories
          in: query
          description: 'IDs das categorias'
          required: false
          schema:
            type: array
            items:
              type: integer
      responses:
        '200':
          description: ...
    delete:
      tags:
        - books
      summary: 'Apaga um livro'
      operationId: DeleteBook
      parameters:
        -
          name: id
          in: path
          description: 'ID do livro'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ...
  /categories:
    get:
      tags:
        - categories
      summary: 'Lista todas as categorias'
      operationId: GetCategories
      parameters:
        -
          name: name
          in: query
          description: 'Nome da categoria'
          required: false
          schema:
            type: string
        -
          name: includes
          in: query
          description: 'Faz o include das relações '
          required: false
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: ...
    post:
      tags:
        - categories
      summary: 'Cria uma nova categoria'
      operationId: PostCategories
      parameters:
        -
          name: name
          in: query
          description: 'Nome da categoria'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ...
  '/categories/{id}':
    get:
      tags:
        - categories
      summary: 'Lista uma categoria'
      operationId: GetCategory
      parameters:
        -
          name: id
          in: path
          description: 'ID da categoria'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ...
    put:
      tags:
        - categories
      summary: 'Atualiza uma categoria'
      operationId: PutCategories
      parameters:
        -
          name: id
          in: path
          description: 'ID da categoria'
          required: true
          schema:
            type: integer
        -
          name: name
          in: query
          description: 'Nome da categoria'
          required: false
          schema:
            type: number
      responses:
        '200':
          description: ...
    delete:
      tags:
        - categories
      summary: 'Apaga uma categoria'
      operationId: DeleteCategory
      parameters:
        -
          name: id
          in: path
          description: 'ID da categoria'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ...
  /rooms:
    get:
      tags:
        - chat
      summary: 'Lista todos os chats do usuário autenticado'
      operationId: GetRooms
      responses:
        '200':
          description: ...
      security:
        -
          apiToken: []
  '/rooms/{room_id}/messages':
    get:
      tags:
        - chat
      summary: 'Lista todas as mensagens de um chat'
      operationId: GetMessages
      parameters:
        -
          name: room_id
          in: path
          description: 'ID do chat'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ...
  '/rooms/{advertiser_id}/w/{buyer_id}':
    post:
      tags:
        - chat
      summary: 'Cria uma mensagem em um chat'
      operationId: PostMessage
      parameters:
        -
          name: advertiser_id
          in: path
          description: 'ID do anunciante'
          required: true
          schema:
            type: integer
        -
          name: buyer_id
          in: path
          description: 'ID do comprador'
          required: true
          schema:
            type: integer
        -
          name: message
          in: query
          description: Mensagem
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ...
      security:
        -
          apiToken: []
...
components:
  securitySchemes:
    apiToken:
      type: apiKey
      description: 'Token do client para consumo da API'
      name: Authorization
      in: header
      scheme: bearer